requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-dialogflow
sqlalchemy
psycopg2-binary
spacy
pyjwt
fr-core-news-md @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.8.0/fr_core_news_md-3.8.0-py3-none-any.whl
